Featurefy exported proptest support
